+++ src/readme.md
@@ -1,13 +1,12 @@
 todo:
 
-2. нажатие на кнопкy физической клавиатуры должно подсвечивать нажатую кнопку на виртуальной
-3. Если нажатие происходит на несколько кнопок, то подсвечиваются все нажатые кнопки (исключений нет и для ctrl-alt-shift)
-4. кнопки для смены языка должны быть указаны на странице c клавиатурой, что бы было понятно, как проверяющему поменять язык на клавиатуре
-5. клики по кнопкам мышкой на виртуальной клавиатуре и нажатие на кнопки физической клавиатуры выводят символы в инпут
-6. нажатие стрелок вниз-вверх-влево-вправо либо выводят соответствующую стрелочку в инпут, либо реализуют навигацию по инпуту
-7. нажатие на enter должно осуществлять перевод каретки
-8. tab создает горизонтальный отступ
-9. при нажатии остальных функциональных клавиш в инпут символы не выводятся
-10. backspace удаляет символ, стоящий перед курсором
-11. del удаляет символ, стоящий после курсора
-12. оставшиеся shift, alt, ctrl, caps lock, space должны работать как в реальной клавиатуре.
+1. нажатие на кнопкy физической клавиатуры должно подсвечивать нажатую кнопку на виртуальной
+2. Если нажатие происходит на несколько кнопок, то подсвечиваются все нажатые кнопки
+3. клики по кнопкам мышкой на виртуальной клавиатуре и нажатие на кнопки физической клавиатуры выводят символы в инпут
+4. нажатие стрелок вниз-вверх-влево-вправо либо выводят соответствующую стрелочку в инпут, либо реализуют навигацию по инпуту
+5. нажатие на enter должно осуществлять перевод каретки
+6. tab создает горизонтальный отступ
+7. при нажатии остальных функциональных клавиш в инпут символы не выводятся
+8. backspace удаляет символ, стоящий перед курсором
+9. del удаляет символ, стоящий после курсора
+10. оставшиеся shift, alt, ctrl, caps lock, space должны работать как в реальной клавиатуре.
